ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
redis==5.0.0
supabase==2.7.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.0
httpx==0.27.0
Pillow==10.4.0
numpy==1.26.4
flower==2.0.1
